解决transform_y、transform_x值设置不正确的BUG。

+++ src/schemas/add_captions.py
@@ -16,8 +16,8 @@ class AddCaptionsRequest(BaseModel):
     line_spacing: Optional[float] = Field(default=None, description="行间距")
     scale_x: float = Field(default=1.0, description="水平缩放")
     scale_y: float = Field(default=1.0, description="垂直缩放")
-    transform_x: int = Field(default=0, description="水平位移")
-    transform_y: int = Field(default=0, description="垂直位移")
+    transform_x: float = Field(default=0.0, description="水平位移")
+    transform_y: float = Field(default=0.0, description="垂直位移")
     style_text: bool = Field(default=False, description="是否使用样式文本")
